Updated user input options
Initially, I wanted the user to have the ability to insert the name of their favorite team in multiple ways. For instance, they could have said "hawks" or "Hawks" or "Atlanta Hawks" or "atlanta hawks" and either of those inputs would have gave them the same responses but I was not able to make it work so I had to narrow it down to one option only.